% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choroplethmapfunc.R
\name{choroplethmap}
\alias{choroplethmap}
\title{choroplethMap}
\usage{
choroplethmap(yeart = "mean")
}
\arguments{
\item{yeart}{an integer year between 2000 and 2017 or "mean", if blank calculates mean}
}
\value{
a chloropleth for the entered year
}
\description{
Creates a choropleth map of fatal encounters for the given year
}
\seealso{
\code{\link{plot_geo}}
}
