% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descstatfuncs.R
\name{descstat}
\alias{descstat}
\title{descStat}
\usage{
descstat(dem, ds = fatalencounters)
}
\arguments{
\item{dem}{a demographic, the demographic the data is to be stratified for (Age, Gender, or Race)}

\item{ds}{a data frame, data frame you want to be stratified, if no input uses fatalencounters}
}
\value{
stratified data frame
}
\description{
Stratifies a data set based on a given demographic
Used within dstable and ds plot
}
\seealso{
\code{\link{dstable}} and \code{\link{dsplot}}
}
