% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descstatfuncs.R
\name{dsplot}
\alias{dsplot}
\title{dsPlot}
\usage{
dsplot(dem)
}
\arguments{
\item{dem}{a demographic, the demographic the data is to be stratified for (Age, Gender, or Race)}
}
\value{
ggplot of the resulting data from an in-method descstat call
}
\description{
Makes a table based off stratified data
}
\seealso{
\code{\link{descstat}} and \code{\link{dstable}} and \code{\link{ggplot2}}
}
