% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descstatfuncs.R
\name{dstable}
\alias{dstable}
\title{dsTable}
\usage{
dstable(dem)
}
\arguments{
\item{dem}{a demographic, the demographic the data is to be stratified for (Age, Gender, or Race)}
}
\value{
data table of the resulting data from an in-method descstat call
}
\description{
Makes a table based off stratified data
}
\seealso{
\code{\link{descstat}} and \code{\link{dsplot}}
}
