% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogramfunc.R
\name{ffcartogram}
\alias{ffcartogram}
\title{ffCartogram}
\usage{
ffcartogram(year)
}
\arguments{
\item{year}{an integer year between 2000 and 2017}
}
\value{
a cartogram for the entered year
}
\description{
Creates a cartogram of fatal encounters for the given year
}
\seealso{
\code{\link{cartogram_cont}}
}
