% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permillcalculation.R
\name{permillcalc}
\alias{permillcalc}
\title{PerMillCalc}
\usage{
permillcalc(x = fatalencounters, capita = TRUE)
}
\arguments{
\item{x}{a data frame of individual fatal encounter cases, default uses fatalencounters}

\item{capita}{a boolean, if TRUE evaluates cases per million population of each state, if FALSE just calculates total cases, default is TRUE}
}
\value{
table a data frame of fatalencounter per state (evaluated per capita or as totals based on capita param)
}
\description{
Calculates the fatal encounters (per million population or total) for each state
}
