% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permillgraphfunc.R
\name{permillgraph}
\alias{permillgraph}
\title{perMillGraph}
\usage{
permillgraph(state, all, capita)
}
\arguments{
\item{state}{a state name, the state that is graphed or highlighted}

\item{all}{a boolean, if TRUE graphs state over the data of other states, if FALSE graphs just the given state}

\item{capita}{a boolean, if TRUE evaluates cases per million population of each state, if FALSE just calculates total cases}
}
\description{
Creates a graph of fatal encounters by state
}
\seealso{
\code{\link{permillcalc}} and \code{\link{permilltable}}
}
