% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permilltablefunc.R
\name{permilltable}
\alias{permilltable}
\title{perMillTable}
\usage{
permilltable(statenm, all, capita)
}
\arguments{
\item{statenm}{a state name, the state that is displayed or highlighted}

\item{all}{a boolean, if TRUE table includes all states, if FALSE table displays just the given state}

\item{capita}{a boolean, if TRUE evaluates cases per million population of each state, if FALSE just calculates total cases}
}
\description{
Creates a table of fatal encounters by state
}
\seealso{
\code{\link{permillcalc}} and \code{\link{permilltable}}
}
